+++ Source/Core/XMLParseTools.cpp
@@ -40,7 +40,7 @@ namespace Core {
 // NOTE: tag *MUST* be in lowercase
 const char* XMLParseTools::FindTag(const char* tag, const char* string, bool closing_tag)
 {
-	int length = strlen(tag);
+	int length = (int)strlen(tag);
 	const char* ptr = string;
 	bool found_closing = false;
